Updated the API and cleaned it up Cleaned up the unused classes
