--- Android/SceneForm/samples/HelloSceneForm/LightingDef.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-
-namespace HelloSceneForm
-{
-    public partial class LightingDef
-    {
-
-    }
-}
