+++ Android/SceneForm/samples/hellosceneform/hellosceneform.csproj
@@ -64,7 +64,6 @@
     <Reference Include="System.Numerics.Vectors" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="LightingDef.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="HelloSceneformActivity.cs" />
