+++ Android/SceneForm/source/Rendering/ViewRenderableBuilder.cs
@@ -1,18 +1,17 @@
-﻿using Android;
-using Android.Runtime;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Android.Runtime;
 
 namespace Google.AR.Sceneform.Rendering
 {
-    public partial class ModelRenderable
+    public partial class ViewRenderable
     {
         public partial class Builder
         {
-            public unsafe ModelRenderable.Builder SetSource(global::Android.Content.Context context, int resource)
+            public unsafe ViewRenderable.Builder SetSource(global::Android.Content.Context context, int resource)
             {
                 var rbuilder = this.JavaCast<Renderable.Builder>();
                 rbuilder.SetSource(context, resource);
@@ -20,7 +19,7 @@ namespace Google.AR.Sceneform.Rendering
                 return this;
             }
 
-            public unsafe ModelRenderable.Builder SetSource(global::Android.Content.Context context, global::Android.Net.Uri sourceUri)
+            public unsafe ViewRenderable.Builder SetSource(global::Android.Content.Context context, global::Android.Net.Uri sourceUri)
             {
                 var rbuilder = this.JavaCast<Renderable.Builder>();
                 rbuilder.SetSource(context, sourceUri);
@@ -28,7 +27,7 @@ namespace Google.AR.Sceneform.Rendering
                 return this;
             }
 
-            public unsafe ModelRenderable.Builder SetSource(global::Android.Content.Context context, global::Java.Util.Concurrent.ICallable inputStreamCreator)
+            public unsafe ViewRenderable.Builder SetSource(global::Android.Content.Context context, global::Java.Util.Concurrent.ICallable inputStreamCreator)
             {
                 var rbuilder = this.JavaCast<Renderable.Builder>();
                 rbuilder.SetSource(context, inputStreamCreator);
@@ -36,7 +35,7 @@ namespace Google.AR.Sceneform.Rendering
                 return this;
             }
 
-            public unsafe ModelRenderable.Builder SetSource(global::Google.AR.Sceneform.Rendering.RenderableDefinition definition)
+            public unsafe ViewRenderable.Builder SetSource(global::Google.AR.Sceneform.Rendering.RenderableDefinition definition)
             {
                 var rbuilder = this.JavaCast<Renderable.Builder>();
                 rbuilder.SetSource(definition);
@@ -44,11 +43,10 @@ namespace Google.AR.Sceneform.Rendering
                 return this;
             }
 
-            public void Build(Action<ModelRenderable> completionAction)
+            public void Build(Action<ViewRenderable> completionAction)
             {
-                this.Build().ThenAccept(new ModelRenderableConsumer(completionAction));
+                this.Build().ThenAccept(new ViewRenderableConsumer(completionAction));
             }
         }
-
     }
 }
